Fixed issues
Issue 1:
Game would improperly end in tie because the checkWinner function
checked to see if there were no space’s left in the board but for
example
x|x|o            x|x|o
o|o|x   —>   o|o|x
_|x|o            o|x|o
That would not be counted as a win because there were no spaces and if
there weren’t then it didn’t check any other cases. I fixed this by
adding the “tie detector” to the end.

Issue 2:
When a computer would win, the computer would wait until the players
next turn to check for a win. I fixed it by making it check for a
winner while it was the computer’s turn.

Issue 3:
We had not made the hard computer function yet, so I made one that
instead of doing what it sounds like it should, it just covers the
whole board, regardless of the players turn.